Fix parsing of unknown commands
This is a proper fix for issue #8

In the last version,  unknown commands would cause parsing failures.
For simplicity we create a `PassWith(s)` object every time we encounter
an unknown command. This might be a speed problem if it happens too
often, and can be worked around at the cost of more complexity if
necessary.